Remove unnecessary "Random" command when signing.
It is not neccessary to invoke the "GenRandom" command before
requesting an ECDSA signature. This command does not generate the
nonce (random value "k") that is required for ECDSA -- that is
performed internally by the ECC508/ECC608 and always has been.

By removing this unnecessary command it should be possible to
generate signatures in less time than before.